ref(platforms): remove platforms from central service module
