+++ app/infrastructure/platforms/__init__.py
@@ -63,7 +63,7 @@ from infrastructure.platforms.providers import (
 from infrastructure.platforms.registry import (
     PlatformRegistry,
 )
-from infrastructure.platforms.service import PlatformService
+from infrastructure.platforms.service import PlatformService, get_platform_service
 
 __all__ = [
     # Service
@@ -113,4 +113,6 @@ __all__ = [
     "AuthenticationError",
     "RateLimitExceededError",
     "FormatterError",
+    # Service
+    "get_platform_service",
 ]
