+++ app/infrastructure/platforms/settings.py
@@ -1,0 +1,9 @@
+from infrastructure.configuration.infrastructure.platforms import (
+    PlatformsSettings,
+    get_platforms_settings,
+)
+
+__all__ = [
+    "PlatformsSettings",
+    "get_platforms_settings",
+]
